# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/user-words/get-words/": {
            "get": {
                "operationId": "user_words_get_words_retrieve",
                "tags": [
                    "user-words"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/user-words/submit-lesson-words/": {
            "post": {
                "operationId": "user_words_submit_lesson_words_create",
                "description": "Dữ liệu nhận từ frontend có cấu trúc:\n{\n    \"is_review\": true,\n    \"lesson_id\": 12,          // Chỉ bắt buộc khi is_review là false\n    \"words\": [\n        {\n            \"word_id\": 123,\n            \"level\": 2,\n            \"streak\": 3,\n            \"is_correct\": true,\n            \"question_type\": \"L2\"\n        },\n        {\n            \"word_id\": 124,\n            \"level\": 3,\n            \"streak\": 2,\n            \"is_correct\": false,\n            \"question_type\": \"L1\"\n        }\n    ]\n}\nLogic xử lý:\n - Nếu is_review = false: đặt new_level = 1, new_streak = 1.\n - Nếu is_review = true:\n       • Nếu is_correct = false: new_streak = 1, new_level = max(level - 1, 1).\n       • Nếu is_correct = true: new_streak = min(10, streak + 1), new_level = min(level + 1, 5).\n - Tính next_review dựa trên new_level, new_streak và question_type.\n - Tạo hoặc cập nhật bản ghi UserWord.\n - Nếu is_review = false, yêu cầu lesson_id để cập nhật UserLesson cho lesson đó.",
                "tags": [
                    "user-words"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}